py-packages/server/pyproject.toml
[tool.poetry]
authors = ["Daven Quinn <dev@davenquinn.com>"]
description = "An API server application for the Corelle plate-rotation system"
license = "MIT"
name = "corelle.server"
version = "1.0.0"

packages = [
  {include = "corelle"},
]

[tool.poetry.dependencies]

Fiona = "^1.8.22"
Flask = "^1"
Flask-RESTful = "^0.3.9"
IPython = "^8"
MarkupSafe = "2.0.1"
PyYAML = "^6.0"
"corelle.client" = {path = "../client", develop = true}
jinja2 = "<3.1.0"
"macrostrat.database" = "^1.0.2"
"macrostrat.utils" = "^1.0.1"
numba = "^0.56.3"
numpy = "^1.23.4"
numpy-quaternion = "^2022.4.2"
psycopg2-binary = "^2.9.5"
python = "^3.8"
rich = "^12.6.0"
simplejson = "^3.17.6"
typer = "^0.6.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
corelle = 'corelle.server.cli:cli'

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
